impelement the extractor for intermediate render
